service and controller work

+++ src/main/java/com/sexyalbum/jdbc/AlbumDao.java
@@ -5,7 +5,7 @@ import com.sexyalbum.model.Album;
 import java.util.List;
 
 public interface AlbumDao {
-    int add(Album album);
+    Long add(Album album);
     int update(Album album);
     int delete(Long albumid);
     Album find(Long albumid);
