+++ src/main/java/com/sexyalbum/jdbc/EleDao.java
@@ -5,7 +5,7 @@ import com.sexyalbum.model.Ele;
 import java.util.List;
 
 public interface EleDao {
-    int add(Ele ele);
+    Long add(Ele ele);
     int update(Ele ele);
     int delete(Long eleid);
     Ele find(Long eleid);
